bugfix to trunk merge

+++ trunk/etc/sglBridge.php
@@ -47,6 +47,9 @@ class SGL_Task_SetupTestDb extends SGL_DecorateProcess
         if ($dbType == 'mysql') {
             $dbType = 'mysql_SGL';
         }
+        if ($dbType == 'mysqli') {
+            $dbType = 'mysqli_SGL';
+        }
     	$protocol = isset($conf['database']['protocol']) ? $conf['database']['protocol'] . '+' : '';
         $dsn = $dbType . '://' .
             $conf['database']['user'] . ':' .
@@ -54,6 +57,9 @@ class SGL_Task_SetupTestDb extends SGL_DecorateProcess
             $protocol .
             $conf['database']['host'];
         $dbh = &SGL_DB::singleton($dsn);
+        if (PEAR::isError($dbh)) {
+            die($dbh->getMessage());
+        }
 
         $query = 'DROP DATABASE IF EXISTS ' . $conf['database']['name'];
         $result = $dbh->query($query);
